+++ SasquatchMac/SasquatchMacObjC/AppDelegate.m
@@ -162,22 +162,22 @@ enum StartupMode { appCenter, oneCollector, both, none, skip };
 #pragma mark - MSACCrashesDelegate
 
 - (BOOL)crashes:(nonnull MSACCrashes *)crashes shouldProcessErrorReport:(nonnull MSACErrorReport *)errorReport {
-  NSLog(@"%@ Should process error report with: %@", kMSLogTag, errorReport.exceptionReason);
+  NSLog(@"%@ Should process error report with description: %@", kMSLogTag, [errorReport description]);
   return YES;
 }
 
 - (void)crashes:(nonnull MSACCrashes *)crashes willSendErrorReport:(nonnull MSACErrorReport *)errorReport {
-  NSLog(@"%@ Will send error report with: %@", kMSLogTag, errorReport.exceptionReason);
+  NSLog(@"%@ Will send error report: %@", kMSLogTag, [errorReport description]);
 }
 
 - (void)crashes:(nonnull MSACCrashes *)crashes didSucceedSendingErrorReport:(nonnull MSACErrorReport *)errorReport {
-  NSLog(@"%@ Did succeed error report sending with: %@", kMSLogTag, errorReport.exceptionReason);
+  NSLog(@"%@ Did succeed sending error report: %@", kMSLogTag, [errorReport description]);
 }
 
 - (void)crashes:(nonnull MSACCrashes *)crashes
     didFailSendingErrorReport:(nonnull MSACErrorReport *)errorReport
                     withError:(nullable NSError *)error {
-  NSLog(@"%@ Did fail sending report with: %@, and error: %@", kMSLogTag, errorReport.exceptionReason, error.localizedDescription);
+  NSLog(@"%@ Did fail sending error report: %@, with error: %@", kMSLogTag, [errorReport description], error.localizedDescription);
 }
 
 - (NSArray<MSACErrorAttachmentLog *> *)attachmentsWithCrashes:(MSACCrashes *)crashes forErrorReport:(MSACErrorReport *)errorReport {