+++ SasquatchTV/SasquatchTVObjC/AppDelegate.m
@@ -97,22 +97,22 @@
 #pragma mark - MSACCrashesDelegate
 
 - (BOOL)crashes:(nonnull MSACCrashes *)crashes shouldProcessErrorReport:(nonnull MSACErrorReport *)errorReport {
-  NSLog(@"Should process error report with: %@", errorReport.exceptionReason);
+  NSLog(@"Should process error report with description: %@", [errorReport description] );
   return YES;
 }
 
 - (void)crashes:(nonnull MSACCrashes *)crashes willSendErrorReport:(nonnull MSACErrorReport *)errorReport {
-  NSLog(@"Will send error report with: %@", errorReport.exceptionReason);
+  NSLog(@"Will send error report: %@", [errorReport description]);
 }
 
 - (void)crashes:(nonnull MSACCrashes *)crashes didSucceedSendingErrorReport:(nonnull MSACErrorReport *)errorReport {
-  NSLog(@"Did succeed error report sending with: %@", errorReport.exceptionReason);
+  NSLog(@"Did succeed sending error report: %@", [errorReport description]);
 }
 
 - (void)crashes:(nonnull MSACCrashes *)crashes
     didFailSendingErrorReport:(nonnull MSACErrorReport *)errorReport
                     withError:(nullable NSError *)error {
-  NSLog(@"Did fail sending report with: %@, and error: %@", errorReport.exceptionReason, error.localizedDescription);
+  NSLog(@"Did fail sending error report: %@, with error: %@", [errorReport description], error.localizedDescription);
 }
 
 - (NSArray<MSACErrorAttachmentLog *> *)attachmentsWithCrashes:(MSACCrashes *)crashes forErrorReport:(MSACErrorReport *)errorReport {