+++ SasquatchTV/SasquatchTVSwift/AppDelegate.swift
@@ -63,16 +63,28 @@ class AppDelegate : UIResponder, UIApplicationDelegate, CrashesDelegate {
 
   // Crashes Delegate
   func crashes(_ crashes: Crashes!, shouldProcess errorReport: ErrorReport!) -> Bool {
+    if errorReport.exceptionReason != nil {
+      NSLog("Should process error report with description: %@", errorReport.description);
+    }
     return true
   }
 
   func crashes(_ crashes: Crashes!, willSend errorReport: ErrorReport!) {
+    if errorReport.exceptionReason != nil {
+      NSLog("Will send error report: %@", errorReport.description);
+    }
   }
 
   func crashes(_ crashes: Crashes!, didSucceedSending errorReport: ErrorReport!) {
+    if errorReport.exceptionReason != nil {
+      NSLog("Did succeed sending error report: %@", errorReport.description);
+    }
   }
 
   func crashes(_ crashes: Crashes!, didFailSending errorReport: ErrorReport!, withError error: Error?) {
+    if errorReport.exceptionReason != nil {
+      NSLog("Did fail sending error report: %@, with error: %@", errorReport.description, error?.localizedDescription ?? "null");
+    }
   }
 
   func attachments(with crashes: Crashes!, for errorReport: ErrorReport!) -> [ErrorAttachmentLog] {